HAWKULAR-1037 update UI to be more aligned Move resource selector and feed selector to the center of page. Move resource tree to it's own component
Remove funcion, we use ES6 coding
