+++ console/src/main/scripts/plugins/metrics/html/directives/explorer/radio-filter.html
@@ -10,7 +10,7 @@
                name="{{config.title}}"
                ng-model="checked"
                ng-checked="false"
-               ng-click="filterClicked({value: value})"> {{value.id}}
+               ng-click="filterClicked({resource: value})"> {{value.id}}
       </label>
     </div>
   </div>
