+++ console/src/main/scripts/plugins/metrics/html/explorer.html
@@ -1,88 +1,53 @@
-<div ng-controller="InventoryExplorerController as exc">
-  <div class="hk-sidebar hk-explorer-sidebar" >
-    <hk-side-filter-container selected-feed="exc.selectedFeed"
-                              selected-resource="exc.selectedResource"
-                              load-metrics="exc.getMetrics(resource)"></hk-side-filter-container>
-  </div>
-  <div class="hk-content-sidebar">
-    <hawkular-subtab ng-controller="Subtab.SubtabController">
-    </hawkular-subtab>
+<div class="hk-explorer" ng-controller="InventoryExplorerController as exc">
 
-      <div class="row">
-          <div class="col-lg-9">
-              <div class="row row-cards-pf">
-                  <div class="col-md-4">
-                     <div class="card-pf">
-                      <h3>Selected Feed</h3>
-                      <span class="hk-selected-resource">{{exc.selectedFeed.id}}</span>
-                     </div>
-                  </div>
-                  <div class="col-md-4">
-                    <div class="card-pf">
-                      <h3>Selected Resource</h3>
-                      <span class="hk-selected-resource">{{exc.selectedResource.id}}</span>
-                    </div>
-                  </div>
-                  <div class="col-md-4">
-                      <div class="card-pf">
-                      <div class="dropdown" dropdown>
-                          <button id="btn-metric" type="button" class="btn btn-primary" dropdown-toggle>
-                              Metric <span class="caret"></span>
-                          </button>
-                          <ul class="dropdown-menu" role="menu" aria-labelledby="btn-append-to-body">
-                              <li ng-repeat="metric in exc.metrics" role="presentation">
-                                  <a role="menutitem" href="#" ng-click="exc.selectMetric(metric)">{{ metric.name }}</a>
-                              </li>
-                          </ul>
-                      </div>
-                      <span class="hk-selected-resource">{{exc.selectedMetric.name}}</span>
-                    </div>
-                  </div>
-              </div>
-          </div>
-          <div class="col-lg-3">
-              <button id="btn-add" type="button" class="btn btn-primary"
-                      ng-click="exc.showChart()" ng-disabled="!exc.buttonActive">
-                  Add to Display
-              </button>
-          </div>
-      </div>
+  <hawkular-subtab ng-controller="Subtab.SubtabController">
+  </hawkular-subtab>
 
-      <div class="row">
-        <div class="col-lg-9">
-          <div class="hk-update pull-right">
-            <button class="btn btn-link hk-chart-update" tooltip-trigger tooltip-placement="top" tooltip="Update chart"
-                    tooltip-append-to-body="true" ng-click="exc.refresh()"><i class="fa fa-refresh"></i></button>
-          </div>
-          <div class="row row-cards-pf">
-            <div class="col-xs-12">
-              <div ng-repeat="chart in exc.charts" class="card-pf hk-graph">
-                <h2>{{chart.name}}</h2>
-                <a href="" ng-click="exc.removeChart(chart)" title="Remove this chart">
-                  <span class="pull-right"><i class="fa fa-trash-o"></i></span>
-                </a>
-                <hawkular-chart
-                  ng-if="exc.chartType[chart.id] !=='AVAILABILITY'"
-                  data="exc.chartData[chart.id]"
-                  chart-type="line"
-                  hide-high-low-values="false"
-                  use-zero-min-value="false"
-                  y-axis-units="{{chart.type.unit}}"
-                  chart-height="250">
-                </hawkular-chart>
-                <availability-chart ng-if="exc.chartType[chart.id] ==='AVAILABILITY'"
-                                    data="exc.chartData[chart.id]"
-                                    chart-height="250">
-                </availability-chart>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="col-lg-3">
-          <hk-alert-info list="exc.alertList" limit="(exc.showAllAlerts ? 100000 : 3)"
-                         resource-id="exc.$routeParams.feedId + '/' + exc.$routeParams.resourceId" title="Server">
-          </hk-alert-info>
-        </div>
+  <div class="row hk-cards-offset">
+    <div class="col-xs-12">
+      <hk-side-filter-container selected-feed="exc.selectedFeed"
+                                selected-resource="exc.selectedResource"
+                                selected-metric="exc.selectedMetric"
+                                on-activate-metric="exc.activateMetric()">
+      </hk-side-filter-container>
     </div>
   </div>
+
+  <div class="row">
+    <div class="col-lg-9 col-md-8">
+      <div class="hk-update pull-right">
+        <button class="btn btn-link hk-chart-update" tooltip-trigger tooltip-placement="top" tooltip="Update chart"
+                tooltip-append-to-body="true" ng-click="exc.refresh()"><i class="fa fa-refresh"></i></button>
+      </div>
+      <div class="row row-cards-pf">
+        <div class="col-xs-12">
+          <div ng-repeat="chart in exc.charts" class="card-pf hk-graph">
+            <h2>{{chart.name}}</h2>
+            <a href="" ng-click="exc.removeChart(chart)" title="Remove this chart">
+              <span class="pull-right"><i class="fa fa-trash-o"></i></span>
+            </a>
+            <hawkular-chart
+              ng-if="exc.chartType[chart.id] !=='AVAILABILITY'"
+              data="exc.chartData[chart.id]"
+              chart-type="line"
+              hide-high-low-values="false"
+              use-zero-min-value="false"
+              y-axis-units="{{chart.type.unit}}"
+              chart-height="250">
+            </hawkular-chart>
+            <availability-chart ng-if="exc.chartType[chart.id] ==='AVAILABILITY'"
+                                data="exc.chartData[chart.id]"
+                                chart-height="250">
+            </availability-chart>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="col-lg-3 col-md-4">
+      <hk-alert-info list="exc.alertList" limit="(exc.showAllAlerts ? 100000 : 3)"
+                     resource-id="exc.$routeParams.feedId + '/' + exc.$routeParams.resourceId" title="Server">
+      </hk-alert-info>
+    </div>
+  </div>
+
 </div>
