+++ console/src/main/scripts/plugins/metrics/less/metrics.less
@@ -2355,6 +2355,8 @@ h2 + .tiles-view-pf {
 
 .hk-urls-list {
 
+  padding: 0 @grid-gutter-width 0 @grid-gutter-width/4;
+
   &.hk-urls-empty {
     margin-top: @grid-gutter-width*4;
   }
@@ -2968,4 +2970,12 @@ section {
 
 .hk-tokens .row-cards-pf {
   margin-top: @grid-gutter-width/2;
+}
+
+// Explorer screen
+
+.hk-resource-tree {
+  padding-top: 10px;
+  border-top: 1px solid @table-border-color;
+  margin-top: 15px;
 }
