+++ console/src/main/scripts/plugins/metrics/ts/directives/explorer/sideFilterContainerDirective.ts
@@ -25,9 +25,16 @@ module HawkularMetrics {
     public selectedResource: any;
     public feedsConfig: any;
     public resourceConfig: any;
+    public selectedMetric: any;
+    public buttonActive: boolean;
+    public metrics: any[] = [];
     public childrenTree: any[] = [];
+    public onResourceSelected: ng.IDeferred<any>;
+    public onActivateMetric: () => void;
+
     /*@ngInject*/
-    constructor(private HawkularInventory: any) {
+    constructor(private HawkularInventory: any, private $q: ng.IQService) {
+      this.onResourceSelected = this.$q.defer();
       this.feedsConfig = {
         title: 'Feed'
       };
@@ -46,7 +53,7 @@ module HawkularMetrics {
     }
 
     public feedSelected(feed) {
-      this.selectedResource = null;
+      this.selectedResource = undefined;
       this.HawkularInventory.ResourceUnderFeed.query({
         feedId: feed.id
       }).$promise.then((resourceList) => {
@@ -57,80 +64,34 @@ module HawkularMetrics {
 
     public resourceSelected(resource) {
       this.onLoadMetrics(resource);
-      this.fetchResourceChildren(resource, this.selectedFeed);
-    }
-
-    private fetchResourceChildren(resource, feed) {
-      this.HawkularInventory.ResourceRecursiveChildrenUnderFeed.get({
-        feedId: feed.id,
-        resourcePath: resource.id.replace(/\//g, '%2F')
-      }).$promise.then((children) => {
-        let preffix = feed.id + InventoryExplorerController.pathDelimiter;
-        let grouppedChildren = _.groupBy(children, (item: any) => item.path.substring(
-          item.path.lastIndexOf(preffix) + preffix.length,
-          item.path.lastIndexOf(InventoryExplorerController.pathDelimiter))
-        );
-        let orderedChildren = {};
-        Object.keys(grouppedChildren).sort().forEach((key) => {
-          orderedChildren[key] = grouppedChildren[key];
-        });
-        let childrenAsTree = _.reduce(orderedChildren, (curr: any, value: any, key: string) => {
-          const path = key.split(InventoryExplorerController.pathDelimiter);
-          if (path.length === 1) {
-            curr.text = SideFilterController.stripSpecialCharsFromLabel(path[0]);
-            curr.nodes = SideFilterController.fillChildrenNodes(value);
-          } else {
-            path.shift();
-            this.insertAtPath(curr, path, value);
-          }
-          return curr;
-        }, _.cloneDeep(this.selectedResource));
-
-        if (children.length !== 0) {
-          this.childrenTree = [childrenAsTree];
-        } else {
-          this.childrenTree.length = 0;
-        }
-      });
-    }
-
-    private static stripSpecialCharsFromLabel(childId): string {
-      let label = childId.substring(childId.lastIndexOf('/') + 1, childId.length);
-      label = (label.lastIndexOf('%2F') !== -1)
-        ? label.substring(label.lastIndexOf('%2F') + '%2F'.length, label.length)
-        : label;
-      return label;
-    }
-
-    private static fillChildrenNodes(children): any[] {
-      _.each(children, (child: any) => {
-        child.text = SideFilterController.stripSpecialCharsFromLabel(child.id);
-      });
-      return children;
-    }
-
-    public insertAtPath(parent: any, path: any, item: any) {
-      if (path.length === 0) {
-        parent.nodes = SideFilterController.fillChildrenNodes(item);
-        return parent;
-      } else {
-        const currId = path[0];
-        path.shift();
-        return this.insertAtPath(
-          _.find(parent.nodes, {text: SideFilterController.stripSpecialCharsFromLabel(currId)}),
-          path, item);
-      }
     }
 
     private onLoadMetrics(resource) {
+      this.selectedResource = resource;
       resource.resourcePath = resource.path.substr(
         resource.path.indexOf(this.selectedFeed.id) + this.selectedFeed.id.length
       );
-      this.selectedResource = resource;
-      if (this.hasOwnProperty('loadMetrics') && typeof this['loadMetrics'] === 'function') {
-        this['loadMetrics']({resource: this.selectedResource});
-      }
+      this.getMetrics(resource);
     }
+
+    public getMetrics(resource) {
+      this.selectedMetric = null;
+      this.buttonActive = false;
+      this.HawkularInventory['MetricOfResourceUnderFeed']['get']({
+          feedId: this.selectedFeed.id,
+          resourcePath: resource.resourcePath.replace(/\/r;/g, '/').slice(1)
+        },
+        (metricList) => {
+          this.metrics = metricList;
+        }
+      );
+    }
+
+    public selectMetric(metric) {
+      this.selectedMetric = metric;
+      this.buttonActive = true;
+    }
+
   }
 
   class HkSideFilterContainerDirective {
@@ -139,7 +100,8 @@ module HawkularMetrics {
     public bindings = {
       selectedResource: '=',
       selectedFeed: '=',
-      loadMetrics: '&'
+      selectedMetric: '=',
+      onActivateMetric: '&'
     };
     /*@ngInject*/
     public controller = SideFilterController;
